Sort out unit tests for organisation API
